fix codecheck

+++ services/audio_engine/node/src/hpae_resample_node.cpp
@@ -118,7 +118,7 @@ HpaePcmBuffer *HpaeResampleNode::SignalProcess(const std::vector<HpaePcmBuffer *
 void HpaeResampleNode::ResampleProcess(float *srcData, uint32_t inputFrameLen, float *dstData, uint32_t outputFrameLen)
 {
     resampler_->Process(srcData, inputFrameLen, dstData, outputFrameLen);
-    int32_t addZeroLen = GetFrameLen() - outputFrameLen > 0 ? GetFrameLen() - outputFrameLen : 0;
+    uint32_t addZeroLen = GetFrameLen() - outputFrameLen > 0 ? GetFrameLen() - outputFrameLen : 0;
 
     if (preNodeInfo_.channels == GetChannelCount()) {
 #ifdef ENABLE_HOOK_PCM
@@ -132,10 +132,10 @@ void HpaeResampleNode::ResampleProcess(float *srcData, uint32_t inputFrameLen, f
     }
     
     float *targetData = resampleOutput_.GetPcmDataBuffer();
-    size_t targetChannels = GetChannelCount();
-    for (int32_t i = 0; i < (int32_t)outputFrameLen; ++i) {
-        for (int32_t ch = 0; ch < (int32_t)targetChannels; ++ch) {
-            size_t leftChIndex = std::min(ch, (preNodeInfo_.channels - 1));
+    uint32_t targetChannels = GetChannelCount();
+    for (uint32_t i = 0; i < outputFrameLen; ++i) {
+        for (uint32_t ch = 0; ch < targetChannels; ++ch) {
+            uint32_t leftChIndex = std::min(ch, (preNodeInfo_.channels - 1));
             if (i < addZeroLen) {
                 targetData[i * targetChannels + ch] = 0;
             } else {
